v0.10.4: Type-checker clean — resolve final 9 Pyright diagnostics
Bring the six core modules to 0 errors / 0 warnings (was 9 errors).
No production-path behavioral change (468 passed, 34 xfail).

- sentence_splitter.py: retype _process_whisper_punctuation -> List[Sentence]
  (stale list[str] annotation since v0.6.0); dead str-fallback builds a Sentence
- punctuation_restorer.py: annotate replace_word(m: re.Match) -> str / w: str
  (flagged re.sub-None was a false positive — .get default is non-None);
  delete dead _domain_safe_split_preserving_delims + unused import
- podscripter.py: diarization_result -> DiarizationResult|None (TYPE_CHECKING
  forward-ref); whisper/merged_boundaries -> list[int]|None (char positions);
  cast segments/speaker_segments at invariance sites; narrow dump guard via local
- speaker_diarization.py: write_diarization_dump boundary params -> list[int]|None

+++ speaker_diarization.py
@@ -330,8 +330,8 @@ def _merge_boundaries(
 def write_diarization_dump(
     diarization_result: DiarizationResult,
     output_file: str,
-    merged_boundaries: list[float] | None = None,
-    whisper_boundaries: list[float] | None = None,
+    merged_boundaries: list[int] | None = None,
+    whisper_boundaries: list[int] | None = None,
 ) -> None:
     """
     Write diarization data to a file for debugging purposes.
